boundDeleteProject

+++ src/main/java/com/example/SimbirSoft_2021/service/ReleaseService.java
@@ -39,12 +39,15 @@ public class ReleaseService implements StandartServiceInterface, ProjectServiceI
     }
 
     @Transactional
-    @Override
+    @Override // ----------------- регистрация
     public ReleaseDto registration(Object o) throws ReleaseExistsException, ReleaseDateFormatException {
         ReleaseDto releaseDto = (ReleaseDto) o;
         DateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss"); // формат времени
         //Calendar cal = Calendar.getInstance(); // вытащить дату из системы
+
+        //  проверка на формат
         try {
+            // присваивание новых значений из форматированных под стандарт данных прешедших значений
             releaseDto.setDataStart(dateFormat.format(dateFormat.parse(releaseDto.getDataStart())));
             releaseDto.setDataEnd(dateFormat.format(dateFormat.parse(releaseDto.getDataEnd())));
         }
@@ -53,23 +56,30 @@ public class ReleaseService implements StandartServiceInterface, ProjectServiceI
         }
 
         ReleaseEntity releaseEntity = ReleaseMapper.INSTANCE.toEntity(releaseDto);
+
         /* // ReleaseExistsException
         if (releaseCrud.findByDataStartAndDataEnd(releaseEntity.getDataStart(), releaseEntity.getDataEnd())!=null){
             throw new ReleaseExistsException();
-        }
-         */
+        }*/
+
+        // сохранение
         releaseCrud.save(releaseEntity);
         return ReleaseMapper.INSTANCE.toDto(releaseEntity);
     }
 
     @Transactional
-    @Override
+    @Override // ----------------- вытащить все реализации
     public List<ReleaseDto> getAll() throws ReleaseNotFoundException {
         List<ReleaseEntity> releaseEntityList = releaseCrud.findAll();
+
+        //  проверка на то что реализации вообще существуют
         if (releaseEntityList==null){
             throw new ReleaseNotFoundException();
         }
+
         List<ReleaseDto> releaseDtoList = new ArrayList<>();
+
+        //  вытаскиваем по одной реализации и сохраняем в коллекцию
         for (ReleaseEntity e:releaseEntityList){
             releaseDtoList.add(ReleaseMapper.INSTANCE.toDto(e));
         }
@@ -77,35 +87,46 @@ public class ReleaseService implements StandartServiceInterface, ProjectServiceI
     }
 
     @Transactional
-    @Override
+    @Override // ----------------- вытащить одну реализацию
     public ReleaseDto getOne(Long id) throws ReleaseNotFoundException {
         ReleaseEntity releaseEntity = releaseCrud.findByReleaseId(id);
+
+        //  проверка на то что реализации вообще существуют
         if (releaseEntity==null){
             throw new ReleaseNotFoundException();
         }
+
         return ReleaseMapper.INSTANCE.toDto(releaseEntity);
     }
 
     @Transactional
-    @Override
+    @Override // ----------------- удалить одну реализацию
     public Long deleteOne(Long id) throws ReleaseNotFoundException {
-        if (releaseCrud.findByReleaseId(id)==null){
-            throw new ReleaseNotFoundException();
+        if (id!=null){
+            //  проверка на то что реализации вообще существуют
+            if (releaseCrud.findByReleaseId(id)==null){
+                throw new ReleaseNotFoundException();
+            }
+            releaseCrud.deleteById(id);
         }
-        releaseCrud.deleteById(id);
         return id;
     }
 
     @Transactional
-    @Override
+    @Override // ----------------- обновить один проект
     public ReleaseDto updateOne(Long id, Object o) throws ReleaseNotFoundException, ReleaseExistsException, ReleaseDateFormatException {
+
+        //  проверка на то что реализации вообще существуют
         if (releaseCrud.findByReleaseId(id)==null){
             throw new ReleaseNotFoundException();
         }
+
         ReleaseDto releaseDto = (ReleaseDto) o;
         DateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss"); // формат времени
         //Calendar cal = Calendar.getInstance(); // вытащить дату из системы
-        try {
+
+        //  проверка на формат
+        try { // присваивание новых значений из форматированных под стандарт данных прешедших значений
             releaseDto.setDataStart(dateFormat.format(dateFormat.parse(releaseDto.getDataStart())));
             releaseDto.setDataEnd(dateFormat.format(dateFormat.parse(releaseDto.getDataEnd())));
         }
@@ -114,19 +135,18 @@ public class ReleaseService implements StandartServiceInterface, ProjectServiceI
         }
 
         ReleaseEntity releaseEntityNew = ReleaseMapper.INSTANCE.toEntity(releaseDto);
-        /* // ReleaseExistsException
-        if (releaseCrud.findByDataStartAndDataEnd(releaseEntity.getDataStart(), releaseEntity.getDataEnd())!=null){
-            throw new ReleaseExistsException();
-        }
-         */
-
         ReleaseEntity releaseEntity = releaseCrud.findByReleaseId(id);
 
+        //  проверка
         if (releaseCrud.findByDataStartAndDataEnd(releaseEntityNew.getDataStart(), releaseEntityNew.getDataEnd())!=null){
             throw new ReleaseExistsException();
         }
+
+        // присваивание новых значений
         releaseEntity.setDataStart(releaseEntityNew.getDataStart());
         releaseEntity.setDataEnd(releaseEntityNew.getDataEnd());
+
+        // сохранение
         releaseCrud.save(releaseEntity);
         return ReleaseMapper.INSTANCE.toDto(releaseEntity);
     }
